stream container logs to backend
