+++ cri/cri.go
@@ -10,21 +10,15 @@ import (
 	"net"
 	"os"
 	"path/filepath"
-	"reflect"
-	"regexp"
-	"sort"
 	"strconv"
 	"strings"
 	"sync"
 	"time"
 
 	"github.com/ddosify/alaz/log"
+	"github.com/fsnotify/fsnotify" //nolint:staticcheck
 
-	// "github.com/gogo/protobuf/jsonpb"
-
-	"github.com/fsnotify/fsnotify"
-	"github.com/golang/protobuf/jsonpb" //nolint:staticcheck
-	"github.com/golang/protobuf/proto"  //nolint:staticcheck
+	//nolint:staticcheck
 	internalapi "k8s.io/cri-api/pkg/apis"
 	pb "k8s.io/cri-api/pkg/apis/runtime/v1"
 	"k8s.io/kubernetes/pkg/kubelet/cri/remote"
@@ -32,20 +26,32 @@ import (
 
 var defaultRuntimeEndpoints = []string{"unix:///proc/1/root/run/containerd/containerd.sock", "unix:///proc/1/root/run/crio/crio.sock", "unix:///proc/1/root/var/run/cri-dockerd.sock"}
 
+type fileReader struct {
+	mu sync.Mutex
+	*bufio.Reader
+}
+
+type containerPodInfo struct {
+	podUid  string
+	podName string
+	podNs   string
+}
+
 type CRITool struct {
-	rs      internalapi.RuntimeService
-	timeout time.Duration
+	rs internalapi.RuntimeService
 
 	connPool *sync.Pool
 	watcher  *fsnotify.Watcher
 
-	logPathToFile          map[string]*bufio.Reader
+	logPathToFile          map[string]*fileReader
 	logPathToContainerMeta map[string]string
+	containerIdToLogPath   map[string]string
 }
 
-func NewCRITool(ctx context.Context, t time.Duration) (*CRITool, error) {
+func NewCRITool(ctx context.Context) (*CRITool, error) {
 	var res internalapi.RuntimeService
 	var err error
+	t := 10 * time.Second
 	for _, endPoint := range defaultRuntimeEndpoints {
 		res, err = remote.NewRemoteRuntimeService(endPoint, t, nil)
 		if err != nil {
@@ -87,8 +93,9 @@ func NewCRITool(ctx context.Context, t time.Duration) (*CRITool, error) {
 		watcher.Close()
 	}()
 
-	logPathToFile := make(map[string]*bufio.Reader, 0)
+	logPathToFile := make(map[string]*fileReader, 0)
 	logPathToContainerMeta := make(map[string]string, 0)
+	containerIdToLogPath := make(map[string]string, 0)
 
 	return &CRITool{
 		rs:                     res,
@@ -96,6 +103,7 @@ func NewCRITool(ctx context.Context, t time.Duration) (*CRITool, error) {
 		watcher:                watcher,
 		logPathToFile:          logPathToFile,
 		logPathToContainerMeta: logPathToContainerMeta,
+		containerIdToLogPath:   containerIdToLogPath,
 	}, nil
 }
 
@@ -139,13 +147,7 @@ func (ct *CRITool) getLogPath(id string) (string, error) {
 	return fmt.Sprintf("/proc/1/root%s", r.Status.LogPath), nil
 }
 
-type ContainerStatusResponse struct {
-	podUid  string
-	podName string
-	podNs   string
-}
-
-func (ct *CRITool) containerStatus(id string) (*ContainerStatusResponse, error) {
+func (ct *CRITool) containerStatus(id string) (*containerPodInfo, error) {
 	if id == "" {
 		return nil, fmt.Errorf("ID cannot be empty")
 	}
@@ -160,9 +162,6 @@ func (ct *CRITool) containerStatus(id string) (*ContainerStatusResponse, error) 
 	info := map[string]interface{}{}
 	json.Unmarshal([]byte(r.Info["info"]), &info)
 
-	// logPath := r.Status.GetLogPath()
-	// pid := info["pid"].(float64)
-
 	sandBoxID := info["sandboxID"].(string)
 
 	podRes, err := ct.rs.PodSandboxStatus(context.TODO(), sandBoxID, verbose)
@@ -173,10 +172,8 @@ func (ct *CRITool) containerStatus(id string) (*ContainerStatusResponse, error) 
 	podUid := podRes.Status.Metadata.Uid
 	podName := podRes.Status.Metadata.Name
 	podNamespace := podRes.Status.Metadata.Namespace
-	// fmt.Printf("containerID:%s\n", id)
-	// fmt.Printf("podUid:%s podName:%s podNs:%s\n", podUid, podName, podNamespace)
 
-	return &ContainerStatusResponse{
+	return &containerPodInfo{
 		podUid:  podUid,
 		podName: podName,
 		podNs:   podNamespace,
@@ -205,7 +202,7 @@ func (ct *CRITool) getContainersOfPod(podSandboxId string) ([]*pb.Container, err
 	return list, nil
 }
 
-func (ct *CRITool) getPods(podUid string) ([]*pb.PodSandbox, error) {
+func (ct *CRITool) getPod(podUid string) ([]*pb.PodSandbox, error) {
 	filter := &pb.PodSandboxFilter{}
 
 	filter.LabelSelector = map[string]string{
@@ -229,8 +226,7 @@ func getContainerMetadataLine(podNs, podName, podUid, containerName string, num 
 	return fmt.Sprintf("**AlazLogs_%s_%s_%s_%s_%d**\n", podNs, podName, podUid, containerName, num)
 }
 
-func (ct *CRITool) readerForLogPath(logPath string) (*bufio.Reader, error) {
-	// TODO: refactor
+func (ct *CRITool) readerForLogPath(logPath string) (*fileReader, error) {
 	if reader, ok := ct.logPathToFile[logPath]; ok {
 		return reader, nil
 	}
@@ -240,10 +236,113 @@ func (ct *CRITool) readerForLogPath(logPath string) (*bufio.Reader, error) {
 		return nil, err
 	}
 
+	file.Seek(0, io.SeekEnd) // seek to end of file
 	reader := bufio.NewReader(file)
-	ct.logPathToFile[logPath] = reader
+	ct.logPathToFile[logPath] = &fileReader{
+		mu:     sync.Mutex{},
+		Reader: reader,
+	}
 
-	return reader, nil
+	return ct.logPathToFile[logPath], nil
+}
+
+func (ct *CRITool) watchContainer(id string, name string) error {
+	logPath, err := ct.getLogPath(id)
+	if err != nil {
+		log.Logger.Error().Err(err).Msgf("Failed to get log path for container %s", id)
+		return err
+	}
+
+	_, err = ct.readerForLogPath(logPath)
+	if err != nil {
+		log.Logger.Error().Err(err).Msgf("Failed to get reader for log path %s", logPath)
+		return err
+	}
+
+	err = ct.watcher.Add(logPath)
+	if err != nil {
+		log.Logger.Error().Err(err).Msgf("Failed to add log path %s to watcher", logPath)
+		return err
+	}
+	ct.containerIdToLogPath[id] = logPath
+
+	fileName := filepath.Base(logPath)
+	fileNameWithoutExt := strings.TrimSuffix(fileName, filepath.Ext(fileName))
+	suffixNum, err := strconv.Atoi(fileNameWithoutExt)
+	if err != nil {
+		log.Logger.Error().Err(err).Msgf("Failed to parse numeric part of log file name %s", fileName)
+	}
+
+	resp, err := ct.containerStatus(id)
+	if err != nil {
+		log.Logger.Error().Err(err).Msgf("Failed to get container status for container %s", id)
+		return err
+	}
+
+	ct.logPathToContainerMeta[logPath] = getContainerMetadataLine(resp.podNs, resp.podName, resp.podUid, name, suffixNum)
+	return nil
+}
+
+func (ct *CRITool) unwatchContainer(id string) {
+	logPath := ct.containerIdToLogPath[id]
+	log.Logger.Info().Msgf("removing container: %s, %s", id, logPath)
+
+	// we must read until EOF and then remove the reader
+	// otherwise the last logs may be lost
+	// trigger manually
+	ct.sendLogs(logPath)
+	log.Logger.Info().Msgf("manually read for last time for %s", logPath)
+
+	ct.watcher.Remove(logPath)
+
+	// close reader
+	if reader, ok := ct.logPathToFile[logPath]; ok {
+		reader.Reset(nil)
+		delete(ct.logPathToFile, logPath)
+	}
+
+	delete(ct.logPathToContainerMeta, logPath)
+	delete(ct.containerIdToLogPath, id)
+}
+
+func (ct *CRITool) sendLogs(logPath string) error {
+	var err error
+	var conn net.Conn = nil
+	for {
+		c := ct.connPool.Get()
+		if c == nil {
+			log.Logger.Error().Msgf("connection is nil, retrying..")
+			continue
+		}
+		conn = c.(net.Conn)
+		break
+	}
+
+	// send metadata first
+	metaLine := ct.logPathToContainerMeta[logPath]
+	_, err = io.Copy(conn, bytes.NewBufferString(metaLine))
+	if err != nil {
+		log.Logger.Error().Err(err).Msgf("metadata could not be sent to backend: %v", err)
+		return err
+	}
+
+	// send logs
+	reader, ok := ct.logPathToFile[logPath]
+	if !ok || reader == nil {
+		log.Logger.Error().Msgf("reader for log path %s is not found", logPath)
+		return err
+	}
+
+	reader.mu.Lock()
+	_, err = io.Copy(conn, reader)
+	if err != nil {
+		log.Logger.Error().Err(err).Msgf("logs could not be sent to backend: %v", err)
+		return err
+	}
+	reader.mu.Unlock()
+
+	ct.connPool.Put(conn)
+	return nil
 }
 
 func (ct *CRITool) StreamLogs() error {
@@ -253,500 +352,119 @@ func (ct *CRITool) StreamLogs() error {
 	}
 
 	for _, c := range containers {
-		logPath, err := ct.getLogPath(c.Id)
+		err := ct.watchContainer(c.Id, c.Metadata.Name)
 		if err != nil {
-			log.Logger.Error().Err(err).Msgf("Failed to get log path for container %s, %s", c.Id, c.Metadata.Name)
-			continue
+			log.Logger.Error().Err(err).Msgf("Failed to watch container %s, %s", c.Id, c.Metadata.Name)
 		}
-		_, err = ct.readerForLogPath(logPath)
-		if err != nil {
-			log.Logger.Error().Err(err).Msgf("Failed to get reader for log path %s", logPath)
-			continue
-		}
-
-		err = ct.watcher.Add(logPath)
-		if err != nil {
-			log.Logger.Error().Err(err).Msgf("Failed to add log path %s to watcher", logPath)
-			continue
-		}
-
-		fileName := filepath.Base(logPath)
-		fileNameWithoutExt := strings.TrimSuffix(fileName, filepath.Ext(fileName))
-		suffixNum, err := strconv.Atoi(fileNameWithoutExt)
-		if err != nil {
-			log.Logger.Error().Err(err).Msgf("Failed to parse numeric part of log file name %s", fileName)
-		}
-
-		// /var/log/pods/ddosify_log-pod-a_e77e90d7-c071-4e08-a467-76771bc399a0/log-container/0.log
-		// TODO: parse logPath, direct metadataline
-
-		resp, err := ct.containerStatus(c.Id)
-		if err != nil {
-			log.Logger.Error().Err(err).Msgf("Failed to get container status for container %s", c.Id)
-			continue
-		}
-
-		ct.logPathToContainerMeta[logPath] = getContainerMetadataLine(resp.podNs, resp.podName, resp.podUid, c.Metadata.Name, suffixNum)
 	}
 
-	// start listening for fsnotify events
+	// listen for new containers
 	go func() {
+		// poll every 10 seconds
 		for {
-			select {
-			case event, ok := <-ct.watcher.Events:
-				if !ok {
-					return
-				}
+			time.Sleep(10 * time.Second)
 
-				if event.Has(fsnotify.Rename) { // logrotate case
-					logPath := event.Name
-					ct.watcher.Add(logPath)
+			containers, err := ct.getAllContainers()
+			if err != nil {
+				log.Logger.Error().Err(err).Msgf("Failed to get all containers")
+				continue
+			}
 
-					// TODO: on an error, and restart event log suffix is incremented
-					// we need to know a way to handle this case
-					// we need know about stopped containers, and new containers
+			// current containers that are being watched
+			currentContainerIds := make(map[string]struct{}, 0)
+			for id, _ := range ct.containerIdToLogPath {
+				currentContainerIds[id] = struct{}{}
+			}
 
-					// at first try it can still not existent
-					// so we need to wait for it to be created
-					for {
-						_, err := os.Stat(logPath)
-						if err == nil {
-							break
-						} else {
-							log.Logger.Info().Msgf("waiting for file to be created on rename: %s", logPath)
-						}
-						time.Sleep(1 * time.Second)
-					}
+			aliveContainers := make(map[string]struct{}, 0)
 
-					logFile, err := os.Open(logPath) // reopen file
-					if err != nil {
-						log.Logger.Error().Err(err).Msgf("Failed to reopen file on rename event: %s", logPath)
-						continue
-					}
-
-					logFile.Seek(0, io.SeekEnd) // seek to end of file
-					ct.logPathToFile[logPath] = bufio.NewReader(logFile)
-
-					log.Logger.Info().Msgf("reopened file for rename: %s", logPath)
+			for _, c := range containers {
+				aliveContainers[c.Id] = struct{}{}
+				if _, ok := currentContainerIds[c.Id]; ok {
 					continue
-				} else if event.Has(fsnotify.Write) {
-					var err error
-					logPath := event.Name
-					conn := ct.connPool.Get()
-
-					// send metadata first
-					metaLine := ct.logPathToContainerMeta[logPath]
-					_, err = io.Copy(conn.(net.Conn), bytes.NewBufferString(metaLine))
+				} else {
+					// new container
+					log.Logger.Info().Msgf("new container found: %s, %s", c.Id, c.Metadata.Name)
+					err := ct.watchContainer(c.Id, c.Metadata.Name)
 					if err != nil {
-						log.Logger.Error().Err(err).Msgf("metadata could not be sent to backend: %v", err)
-						log.Logger.Error().Msgf("TODO: handle this error, backend may be down, connection may be closed for any reason, etc.")
-						continue
+						log.Logger.Error().Err(err).Msgf("Failed to watch new container %s, %s", c.Id, c.Metadata.Name)
 					}
-
-					// send logs
-					reader, ok := ct.logPathToFile[logPath]
-					if !ok || reader == nil {
-						log.Logger.Error().Msgf("reader for log path %s is not found", logPath)
-						continue
-					}
-
-					n, err := io.Copy(conn.(net.Conn), reader)
-					if err != nil {
-						log.Logger.Error().Err(err).Msgf("logs could not be sent to backend: %v", err)
-						log.Logger.Error().Msgf("TODO: handle this error, backend may be down, connection may be closed for any reason, etc.")
-						// TODO: try to reconnect
-					}
-					log.Logger.Debug().Msgf("written %d bytes to conn for %s", logPath)
-					ct.connPool.Put(conn)
 				}
-			case err, ok := <-ct.watcher.Errors:
-				if !ok {
-					return
+			}
+
+			for id := range currentContainerIds {
+				if _, ok := aliveContainers[id]; !ok {
+					// container is gone
+					ct.unwatchContainer(id)
 				}
-				log.Logger.Error().Err(err).Msgf("fsnotify error")
 			}
 		}
 	}()
 
-	return nil
-}
+	// start listening for fsnotify events
+	go func() {
+		worker := 50
+		// start workers
+		for i := 0; i < worker; i++ {
+			go func() {
+				for {
+					select {
+					case event, ok := <-ct.watcher.Events:
+						if !ok {
+							return
+						}
 
-type listOptions struct {
-	// id of container or sandbox
-	id string
-	// podID of container
-	podID string
-	// Regular expression pattern to match pod or container
-	nameRegexp string
-	// Regular expression pattern to match the pod namespace
-	podNamespaceRegexp string
-	// state of the sandbox
-	state string
-	// show verbose info for the sandbox
-	verbose bool
-	// labels are selectors for the sandbox
-	labels map[string]string
-	// quiet is for listing just container/sandbox/image IDs
-	quiet bool
-	// output format
-	output string
-	// all containers
-	all bool
-	// latest container
-	latest bool
-	// last n containers
-	last int
-	// out with truncating the id
-	noTrunc bool
-	// image used by the container
-	image string
-	// resolve image path
-	resolveImagePath bool
-}
+						if event.Has(fsnotify.Rename) { // logrotate case
+							logPath := event.Name
+							// containerd compresses logs, and recreates the file, it comes as a rename event
+							for {
+								_, err := os.Stat(logPath)
+								if err == nil {
+									break
+								} else {
+									log.Logger.Info().Msgf("waiting for file to be created on rename: %s", logPath)
+								}
+								time.Sleep(1 * time.Second)
+							}
 
-func ListPodSandboxesV2(client internalapi.RuntimeService, opts listOptions) error {
-	filter := &pb.PodSandboxFilter{}
-	if opts.id != "" {
-		filter.Id = opts.id
-	}
-	if opts.state != "" {
-		st := &pb.PodSandboxStateValue{}
-		st.State = pb.PodSandboxState_SANDBOX_NOTREADY
-		switch strings.ToLower(opts.state) {
-		case "ready":
-			st.State = pb.PodSandboxState_SANDBOX_READY
-			filter.State = st
-		case "notready":
-			st.State = pb.PodSandboxState_SANDBOX_NOTREADY
-			filter.State = st
-		default:
-			log.Fatalf("--state should be ready or notready")
+							logFile, err := os.Open(logPath) // reopen file
+							if err != nil {
+								log.Logger.Error().Err(err).Msgf("Failed to reopen file on rename event: %s", logPath)
+								continue
+							}
+
+							err = ct.watcher.Add(logPath)
+							if err != nil {
+								log.Logger.Error().Err(err).Msgf("Failed to add log path %s to watcher", logPath)
+								continue
+							}
+							logFile.Seek(0, io.SeekEnd) // seek to end of file
+							ct.logPathToFile[logPath] = &fileReader{
+								mu:     sync.Mutex{},
+								Reader: bufio.NewReader(logFile),
+							}
+
+							log.Logger.Info().Msgf("reopened file for rename: %s", logPath)
+							continue
+						} else if event.Has(fsnotify.Write) {
+							// TODO: apps that writes too much logs might block small applications and causes lag on small apps logs
+							// we don't have to read logs on every write event ??
+
+							err := ct.sendLogs(event.Name)
+							if err != nil {
+								log.Logger.Error().Err(err).Msgf("Failed to send logs for %s", event.Name)
+							}
+						}
+					case err, ok := <-ct.watcher.Errors:
+						if !ok {
+							return
+						}
+						log.Logger.Error().Err(err).Msgf("fsnotify error")
+					}
+				}
+			}()
 		}
-	}
-	if opts.labels != nil {
-		filter.LabelSelector = opts.labels
-	}
-	// request := &pb.ListPodSandboxRequest{
-	// 	Filter: filter,
-	// }
-	// logrus.Debugf("ListPodSandboxRequest: %v", request)
-	r, err := client.ListPodSandbox(context.TODO(), filter)
-	// logrus.Debugf("ListPodSandboxResponse: %v", r)
-	if err != nil {
-		return err
-	}
-	r = getSandboxesList(r, opts)
-
-	switch opts.output {
-	case "json":
-		return outputProtobufObjAsJSON(&pb.ListPodSandboxResponse{Items: r})
-	case "table":
-	// continue; output will be generated after the switch block ends.
-	default:
-		return fmt.Errorf("unsupported output format %q", opts.output)
-	}
+	}()
 
 	return nil
-}
-
-type sandboxByCreated []*pb.PodSandbox
-
-func (a sandboxByCreated) Len() int      { return len(a) }
-func (a sandboxByCreated) Swap(i, j int) { a[i], a[j] = a[j], a[i] }
-func (a sandboxByCreated) Less(i, j int) bool {
-	return a[i].CreatedAt > a[j].CreatedAt
-}
-
-type containerByCreated []*pb.Container
-
-func (a containerByCreated) Len() int      { return len(a) }
-func (a containerByCreated) Swap(i, j int) { a[i], a[j] = a[j], a[i] }
-func (a containerByCreated) Less(i, j int) bool {
-	return a[i].CreatedAt > a[j].CreatedAt
-}
-
-func getSandboxesList(sandboxesList []*pb.PodSandbox, opts listOptions) []*pb.PodSandbox {
-	filtered := []*pb.PodSandbox{}
-	for _, p := range sandboxesList {
-		// Filter by pod name/namespace regular expressions.
-		if matchesRegex(opts.nameRegexp, p.Metadata.Name) &&
-			matchesRegex(opts.podNamespaceRegexp, p.Metadata.Namespace) {
-			filtered = append(filtered, p)
-		}
-	}
-
-	sort.Sort(sandboxByCreated(filtered))
-	n := len(filtered)
-	if opts.latest {
-		n = 1
-	}
-	if opts.last > 0 {
-		n = opts.last
-	}
-	n = func(a, b int) int {
-		if a < b {
-			return a
-		}
-		return b
-	}(n, len(filtered))
-
-	return filtered[:n]
-}
-
-func outputProtobufObjAsJSON(obj proto.Message) error {
-	marshaledJSON, err := protobufObjectToJSON(obj)
-	if err != nil {
-		fmt.Println(err)
-		return err
-	}
-	fmt.Println("here")
-	fmt.Println(marshaledJSON)
-	return nil
-}
-
-func protobufObjectToJSON(obj proto.Message) (string, error) {
-	jsonpbMarshaler := jsonpb.Marshaler{EmitDefaults: true, Indent: "  "}
-	marshaledJSON, err := jsonpbMarshaler.MarshalToString(obj)
-	if err != nil {
-		return "", err
-	}
-	return marshaledJSON, nil
-}
-
-func matchesRegex(pattern, target string) bool {
-	if pattern == "" {
-		return true
-	}
-	matched, err := regexp.MatchString(pattern, target)
-	if err != nil {
-		// Assume it's not a match if an error occurs.
-		return false
-	}
-	return matched
-}
-
-func getContainersList(containersList []*pb.Container, opts listOptions) []*pb.Container {
-	filtered := []*pb.Container{}
-	for _, c := range containersList {
-		// Filter by pod name/namespace regular expressions.
-		if matchesRegex(opts.nameRegexp, c.Metadata.Name) {
-			filtered = append(filtered, c)
-		}
-	}
-
-	sort.Sort(containerByCreated(filtered))
-	n := len(filtered)
-	if opts.latest {
-		n = 1
-	}
-	if opts.last > 0 {
-		n = opts.last
-	}
-	n = func(a, b int) int {
-		if a < b {
-			return a
-		}
-		return b
-	}(n, len(filtered))
-
-	return filtered[:n]
-}
-
-func ListContainers(runtimeClient internalapi.RuntimeService, opts listOptions) error {
-	filter := &pb.ContainerFilter{}
-	if opts.id != "" {
-		filter.Id = opts.id
-	}
-	if opts.podID != "" {
-		filter.PodSandboxId = opts.podID
-	}
-	st := &pb.ContainerStateValue{}
-	if !opts.all && opts.state == "" {
-		st.State = pb.ContainerState_CONTAINER_RUNNING
-		filter.State = st
-	}
-	if opts.state != "" {
-		st.State = pb.ContainerState_CONTAINER_UNKNOWN
-		switch strings.ToLower(opts.state) {
-		case "created":
-			st.State = pb.ContainerState_CONTAINER_CREATED
-			filter.State = st
-		case "running":
-			st.State = pb.ContainerState_CONTAINER_RUNNING
-			filter.State = st
-		case "exited":
-			st.State = pb.ContainerState_CONTAINER_EXITED
-			filter.State = st
-		case "unknown":
-			st.State = pb.ContainerState_CONTAINER_UNKNOWN
-			filter.State = st
-		default:
-			log.Fatalf("--state should be one of created, running, exited or unknown")
-		}
-	}
-	if opts.latest || opts.last > 0 {
-		// Do not filter by state if latest/last is specified.
-		filter.State = nil
-	}
-	if opts.labels != nil {
-		filter.LabelSelector = opts.labels
-	}
-
-	r, err := runtimeClient.ListContainers(context.TODO(), filter)
-	// logrus.Debugf("ListContainerResponse: %v", r)
-	if err != nil {
-		return err
-	}
-	r = getContainersList(r, opts)
-
-	switch opts.output {
-	case "json":
-		return outputProtobufObjAsJSON(&pb.ListContainersResponse{Containers: r})
-	case "table":
-	// continue; output will be generated after the switch block ends.
-	default:
-		return fmt.Errorf("unsupported output format %q", opts.output)
-	}
-
-	return nil
-}
-
-// ContainerStatus sends a ContainerStatusRequest to the server, and parses
-// the returned ContainerStatusResponse.
-func ContainerStatus(client internalapi.RuntimeService, id, output string, tmplStr string, quiet bool) error {
-	verbose := !(quiet)
-	if output == "" { // default to json output
-		output = "json"
-	}
-	if id == "" {
-		return fmt.Errorf("ID cannot be empty")
-	}
-	// request := &pb.ContainerStatusRequest{
-	// 	ContainerId: id,
-	// 	Verbose:     verbose,
-	// }
-
-	// logrus.Debugf("ContainerStatusRequest: %v", request)
-	r, err := client.ContainerStatus(context.TODO(), id, verbose)
-	// logrus.Debugf("ContainerStatusResponse: %v", r)
-	if err != nil {
-		return err
-	}
-
-	status, err := marshalContainerStatus(r.Status)
-	if err != nil {
-		return err
-	}
-
-	switch output {
-	case "json", "go-template":
-		return outputStatusInfo(status, r.Info, output, tmplStr)
-	// case "table": // table output is after this switch block
-	default:
-		return fmt.Errorf("output option cannot be %s", output)
-	}
-
-	return nil
-}
-
-// marshalContainerStatus converts container status into string and converts
-// the timestamps into readable format.
-func marshalContainerStatus(cs *pb.ContainerStatus) (string, error) {
-	statusStr, err := protobufObjectToJSON(cs)
-	if err != nil {
-		return "", err
-	}
-	jsonMap := make(map[string]interface{})
-	err = json.Unmarshal([]byte(statusStr), &jsonMap)
-	if err != nil {
-		return "", err
-	}
-
-	jsonMap["createdAt"] = time.Unix(0, cs.CreatedAt).Format(time.RFC3339Nano)
-	var startedAt, finishedAt time.Time
-	if cs.State != pb.ContainerState_CONTAINER_CREATED {
-		// If container is not in the created state, we have tried and
-		// started the container. Set the startedAt.
-		startedAt = time.Unix(0, cs.StartedAt)
-	}
-	if cs.State == pb.ContainerState_CONTAINER_EXITED ||
-		(cs.State == pb.ContainerState_CONTAINER_UNKNOWN && cs.FinishedAt > 0) {
-		// If container is in the exit state, set the finishedAt.
-		// Or if container is in the unknown state and FinishedAt > 0, set the finishedAt
-		finishedAt = time.Unix(0, cs.FinishedAt)
-	}
-	jsonMap["startedAt"] = startedAt.Format(time.RFC3339Nano)
-	jsonMap["finishedAt"] = finishedAt.Format(time.RFC3339Nano)
-	return marshalMapInOrder(jsonMap, *cs)
-}
-
-func outputStatusInfo(status string, info map[string]string, format string, tmplStr string) error {
-	// Sort all keys
-	keys := []string{}
-	for k := range info {
-		keys = append(keys, k)
-	}
-	sort.Strings(keys)
-
-	jsonInfo := "{" + "\"status\":" + status + ","
-	for _, k := range keys {
-		var res interface{}
-		// We attempt to convert key into JSON if possible else use it directly
-		if err := json.Unmarshal([]byte(info[k]), &res); err != nil {
-			jsonInfo += "\"" + k + "\"" + ":" + "\"" + info[k] + "\","
-		} else {
-			jsonInfo += "\"" + k + "\"" + ":" + info[k] + ","
-		}
-	}
-	jsonInfo = jsonInfo[:len(jsonInfo)-1]
-	jsonInfo += "}"
-
-	switch format {
-	case "json":
-		var output bytes.Buffer
-		if err := json.Indent(&output, []byte(jsonInfo), "", "  "); err != nil {
-			return err
-		}
-		fmt.Println(output.String())
-	default:
-		fmt.Printf("Don't support %q format\n", format)
-	}
-	return nil
-}
-
-// marshalMapInOrder marshalls a map into json in the order of the original
-// data structure.
-func marshalMapInOrder(m map[string]interface{}, t interface{}) (string, error) {
-	s := "{"
-	v := reflect.ValueOf(t)
-	for i := 0; i < v.Type().NumField(); i++ {
-		field := jsonFieldFromTag(v.Type().Field(i).Tag)
-		if field == "" || field == "-" {
-			continue
-		}
-		value, err := json.Marshal(m[field])
-		if err != nil {
-			return "", err
-		}
-		s += fmt.Sprintf("%q:%s,", field, value)
-	}
-	s = s[:len(s)-1]
-	s += "}"
-	var buf bytes.Buffer
-	if err := json.Indent(&buf, []byte(s), "", "  "); err != nil {
-		return "", err
-	}
-	return buf.String(), nil
-}
-
-// jsonFieldFromTag gets json field name from field tag.
-func jsonFieldFromTag(tag reflect.StructTag) string {
-	field := strings.Split(tag.Get("json"), ",")[0]
-	for _, f := range strings.Split(tag.Get("protobuf"), ",") {
-		if !strings.HasPrefix(f, "json=") {
-			continue
-		}
-		field = strings.TrimPrefix(f, "json=")
-	}
-	return field
 }
